+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add `forget` command to remove the note from server and cache 
+
 ## [0.1.0] - 2026-05-06
 
 - Initial release